+++ src/main/java/roomescape/application/RoomescapeApplication.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.application;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
