+++ src/main/java/roomescape/exeption/GlobalExeptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exeption;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
