+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,5 +1,5 @@
 
-package roomescape.reservation;
+package roomescape.reservation.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
